requirements.txt
numpy
matplotlib
tensorflow
keras
Pillow
streamlit
scipy
scikit-learn